+++ src/main/java/com/equipeor/isepu/model/Subject.java
@@ -9,7 +9,7 @@ public class Subject {
     @Id
     @GeneratedValue
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -23,7 +23,7 @@ public class Subject {
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return this.id;
     }
